requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
cycler==0.10.0
flake8==3.6.0
future==0.17.1
idna==2.7
isort==4.3.4
kiwisolver==1.0.1
langdetect==1.0.7
lazy-object-proxy==1.3.1
matplotlib==3.0.2
mccabe==0.6.1
numpy==1.15.4
oauthlib==2.1.0
pandas==0.23.4
pbr==5.1.1
pycodestyle==2.4.0
pyflakes==2.0.0
pyparsing==2.3.0
PySocks==1.6.8
python-dateutil==2.7.5
python-dotenv==0.10.1
python-twitter==3.5
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.11.0
urllib3==1.24.1
wrapt==1.10.11
